+++ dshop_api/src/main/java/com/nashtech/dshop_api/data/entities/StatusType.java
@@ -17,7 +17,7 @@ import lombok.Setter;
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class StatusType extends AuditEntity<Integer>{
+public class StatusType extends AuditEntity<Long>{
 
     @Column(name = "STATUS_NAME", unique = true)
     private String statusName;
